feat(acp): add Qwen Code ACP support and share ACP client base
## Implementation
- extract the Copilot ACP implementation into BaseACPClient for shared ACP protocol handling
- add QwenACPClient and the qwen-code-acp provider for qwen --acp integration
- wire qwen-code-acp through provider resolution, auth, runtime selection, and model setup flows
- add default ACP base URL constants and document the new environment variable
- include QWEN_CODE_ACP_BASE_URL in provider-related test environment coverage

## WHY
- allow Hermes to use Qwen Code ACP as an external-process provider through the existing OpenAI-compatible flow
- reduce duplicated ACP client logic so future protocol fixes and features can be maintained in one place

## WHY NOT
- copying the Copilot ACP client would increase protocol duplication and make future fixes harder to keep consistent
- folding Qwen Code ACP into an existing provider would blur provider-specific command and base URL behavior instead of keeping the integration explicit
